+++ CHANGELOG.md
@@ -11,10 +11,13 @@ Note: `Chip` was available before but not documented. If you used `Chip` in dmc<
 - Added new and missing props to the charts #349 by @AnnMarieW
 
 
+
 ### Fixed
 
 - Excluded the `loading_state` prop from being passed to the DOM.   Added `data-dash-is-loading` attribute to
 components during callback execution, allowing custom CSS styling for loading states. #325 by @AnnMarieW
+- Enabled the `opened` prop to trigger callback #353 by @AnnMarieW
+
 
 ### Changed